Clearing %ENV to prevent tainted data from crashing the script
The script runs with "taint" enabled.  An explicit $ENV{'PATH'} is set, however
the remaining values in %ENV were untouched.  NRPE 3.2 sets a value for
$ENV{'IFS'} which is used by perl when running an external script/program. This
change in NRPE caused the script to die do to tainted data.  In order to
prevent this issue in the future, all values in %ENV are cleared and then
explicit values are set.